refactor(republik-crowdfundings): Move fn to lib/Errors
