+++ packages/republik-crowdfundings/lib/payments/stripe/Errors.js
@@ -10,4 +10,30 @@ class RequiresPaymentMethodError extends Error {
   }
 }
 
+const throwError = (e, { pledgeId, membershipId, t, kind }) => {
+  if (e.type === 'StripeCardError') {
+    const translatedError = t('api/pay/stripe/' + e.code)
+    if (translatedError) {
+      console.warn(e, { pledgeId, membershipId, kind })
+      throw new Error(translatedError)
+    } else {
+      console.warn('translation not found for stripe error', {
+        pledgeId,
+        kind,
+        e,
+      })
+      throw new Error(e.message)
+    }
+  }
+
+  if (e.name === 'RequiresPaymentMethodError') {
+    console.warn(e, { pledgeId, membershipId, kind })
+    throw new Error(t('api/pay/stripe/card_declined'))
+  }
+
+  console.error(e, { pledgeId, membershipId, kind })
+  throw new Error(t('api/unexpected'))
+}
+
 module.exports.RequiresPaymentMethodError = RequiresPaymentMethodError
+module.exports.throwError = throwError
